Se modifico la estructura del trabajo segun lo visto en clase - se agregaron nuevas carpetas y se reorganizaron algunos archivos. Se agregaron clases
