+++ TFI-GrupoAL/package.json
@@ -2,10 +2,10 @@
   "name": "tfi-grupoal_v1",
   "version": "1.0.0",
   "description": "trabajo final integrador ",
-  "main": "index.js",
+  "main": "src/reservas.js",
   "type": "module",
   "scripts": {
-    "start": "node index.js",
+    "start": "node src/reservas.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
@@ -14,6 +14,6 @@
     "dotenv": "^17.2.2",
     "express": "^5.1.0",
     "express-handlebars": "^8.0.3",
-    "mysql2": "^3.14.5"
+    "mysql2": "^3.15.1"
   }
 }
