+++ TFI-GrupoAL/src/reservas.js
@@ -3,11 +3,13 @@ import expressHandlebars from 'express-handlebars';
 import dotenv from 'dotenv';
 import { fileURLToPath } from 'url';
 import { dirname, join } from 'path';
-import { inicializarAdmin } from './src/utils/inicializarAdmin.js'; 
+// import { inicializarAdmin } from './utils/inicializarAdmin.js'; 
 
 // import { getTodosLosSalones } from './database/conexion.js';
-import salonesRouter from './src/routes/salonesRouter.js';
-import usuariosRouter from './src/routes/usuariosRouter.js';
+// import salonesRouter from './routes/salonesRouter.js';
+// import usuariosRouter from './routes/usuariosRouter.js';
+
+import { router as v1SalonesRutas } from './v1/routes/salonesRouter.js';
 
 dotenv.config();
 
@@ -55,8 +57,8 @@ app.use(express.static(__dirname + '/public'));
 //   res.render('contacto', { title: 'Contacto' });
 // });
 
-app.use('/usuarios', usuariosRouter)
-app.use('/salones', salonesRouter);
+// app.use('/usuarios', usuariosRouter)
+app.use('/api/v1/salones', v1SalonesRutas);
 
 // Página personalizada de error 404
 // app.use((req, res) => {
@@ -72,11 +74,11 @@ app.use('/salones', salonesRouter);
 // });
 
 // CREACION DE ADMIN INICIAL
-inicializarAdmin().then(() => {
-  console.log("Chequeo de admin completo");
-}).catch(err => {
-  console.error("Error al chequear admin:", err);
-});
+// inicializarAdmin().then(() => {
+//   console.log("Chequeo de admin completo");
+// }).catch(err => {
+//   console.error("Error al chequear admin:", err);
+// });
 
 app.listen(port, () =>
   console.log(`Express iniciado en http://localhost:${port}`)
