--- TFI-GrupoAL/src/routes/usuariosRouter.js
@@ -1,11 +1,0 @@
-import { Router } from "express";
-import { listarUsuarios, listarUsuarioPorId, crearUsuario } from "../controllers/usuarios.js";
-
-const router = Router();
-
-//Listar usuarios
-router.get('/', listarUsuarios);
-router.get('/:id', listarUsuarioPorId);
-router.post('/', crearUsuario);
-
-export default router;
