+++ TFI-GrupoAL/src/v1/routes/usuariosRouter.js
@@ -1,0 +1,11 @@
+// import { Router } from "express";
+// import { listarUsuarios, listarUsuarioPorId, crearUsuario } from "../../controllers/usuarios.js";
+
+// const router = Router();
+
+// //Listar usuarios
+// router.get('/', listarUsuarios);
+// router.get('/:id', listarUsuarioPorId);
+// router.post('/', crearUsuario);
+
+// export default router;
